requirements.txt
# Core Requirements
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-openai>=0.0.5

# OpenAI Integration
openai>=1.3.0

# Data Structures
pydantic>=2.0.0
typing-extensions>=4.8.0

# Configuration
python-dotenv>=1.0.0

# Text Processing (for argument validation)
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization (optional but recommended)
# pygraphviz>=1.10(conda install -c conda-forge pygraphviz)
graphviz>=0.20.0

# CLI Enhancement
rich>=13.0.0